requirements.txt
annotated-types==0.6.0
google-generativeai==0.5.2
grpcio-status==1.62.2
psycopg2-binary==2.9.9
pydantic_core==2.18.2
python-dotenv==1.0.1
reflex-magic-link-auth==0.0.3
